refactor(s2n-quic-platform): split out tests and builders
