+++ quic/s2n-quic-platform/src/io/tokio.rs
@@ -17,9 +17,13 @@ use s2n_quic_core::{
 use std::{convert::TryInto, io, io::ErrorKind};
 use tokio::{net::UdpSocket, runtime::Handle};
 
-pub type PathHandle = socket::Handle;
-
+mod builder;
 mod clock;
+#[cfg(test)]
+mod tests;
+
+pub type PathHandle = socket::Handle;
+pub use builder::Builder;
 pub(crate) use clock::Clock;
 
 impl crate::socket::std::Socket for UdpSocket {
@@ -217,119 +221,6 @@ impl Io {
     }
 }
 
-#[derive(Debug, Default)]
-pub struct Builder {
-    handle: Option<Handle>,
-    rx_socket: Option<socket2::Socket>,
-    tx_socket: Option<socket2::Socket>,
-    recv_addr: Option<std::net::SocketAddr>,
-    send_addr: Option<std::net::SocketAddr>,
-    recv_buffer_size: Option<usize>,
-    send_buffer_size: Option<usize>,
-    max_mtu: MaxMtu,
-    max_segments: gso::MaxSegments,
-    reuse_port: bool,
-}
-
-impl Builder {
-    #[must_use]
-    pub fn with_handle(mut self, handle: Handle) -> Self {
-        self.handle = Some(handle);
-        self
-    }
-
-    /// Sets the local address for the runtime to listen on. If no send address
-    /// or tx socket is specified, this address will also be used for transmitting from.
-    ///
-    /// NOTE: this method is mutually exclusive with `with_rx_socket`
-    pub fn with_receive_address(mut self, addr: std::net::SocketAddr) -> io::Result<Self> {
-        debug_assert!(self.rx_socket.is_none(), "rx socket has already been set");
-        self.recv_addr = Some(addr);
-        Ok(self)
-    }
-
-    /// Sets the local address for the runtime to transmit from. If no send address
-    /// or tx socket is specified, the receive_address will be used for transmitting.
-    ///
-    /// NOTE: this method is mutually exclusive with `with_tx_socket`
-    pub fn with_send_address(mut self, addr: std::net::SocketAddr) -> io::Result<Self> {
-        debug_assert!(self.tx_socket.is_none(), "tx socket has already been set");
-        self.send_addr = Some(addr);
-        Ok(self)
-    }
-
-    /// Sets the socket used for receiving for the runtime. If no tx_socket or send address is
-    /// specified, this socket will be used for transmitting.
-    ///
-    /// NOTE: this method is mutually exclusive with `with_receive_address`
-    pub fn with_rx_socket(mut self, socket: std::net::UdpSocket) -> io::Result<Self> {
-        debug_assert!(
-            self.recv_addr.is_none(),
-            "recv address has already been set"
-        );
-        self.rx_socket = Some(socket.into());
-        Ok(self)
-    }
-
-    /// Sets the socket used for transmitting on for the runtime. If no tx_socket or send address is
-    /// specified, the rx_socket will be used for transmitting.
-    ///
-    /// NOTE: this method is mutually exclusive with `with_send_address`
-    pub fn with_tx_socket(mut self, socket: std::net::UdpSocket) -> io::Result<Self> {
-        debug_assert!(
-            self.send_addr.is_none(),
-            "send address has already been set"
-        );
-        self.tx_socket = Some(socket.into());
-        Ok(self)
-    }
-
-    /// Sets the size of the operating system’s send buffer associated with the tx socket
-    pub fn with_send_buffer_size(mut self, send_buffer_size: usize) -> io::Result<Self> {
-        self.send_buffer_size = Some(send_buffer_size);
-        Ok(self)
-    }
-
-    /// Sets the size of the operating system’s receive buffer associated with the rx socket
-    pub fn with_recv_buffer_size(mut self, recv_buffer_size: usize) -> io::Result<Self> {
-        self.recv_buffer_size = Some(recv_buffer_size);
-        Ok(self)
-    }
-
-    /// Sets the largest maximum transmission unit (MTU) that can be sent on a path
-    pub fn with_max_mtu(mut self, max_mtu: u16) -> io::Result<Self> {
-        self.max_mtu = max_mtu
-            .try_into()
-            .map_err(|err| io::Error::new(ErrorKind::InvalidInput, format!("{err}")))?;
-        Ok(self)
-    }
-
-    /// Disables Generic Segmentation Offload (GSO)
-    ///
-    /// By default, GSO will be used unless the platform does not support it or an attempt to use
-    /// GSO fails. If it is known that GSO is not available, set this option to explicitly disable it.
-    pub fn with_gso_disabled(mut self) -> io::Result<Self> {
-        self.max_segments = 1.try_into().expect("1 is always a valid MaxSegments value");
-        Ok(self)
-    }
-
-    /// Enables the port reuse (SO_REUSEPORT) socket option
-    pub fn with_reuse_port(mut self) -> io::Result<Self> {
-        if !cfg!(unix) {
-            return Err(io::Error::new(
-                io::ErrorKind::InvalidInput,
-                "reuse_port is not supported on the current platform",
-            ));
-        }
-        self.reuse_port = true;
-        Ok(self)
-    }
-
-    pub fn build(self) -> io::Result<Io> {
-        Ok(Io { builder: self })
-    }
-}
-
 #[derive(Debug)]
 struct Instance<E> {
     clock: Clock,
@@ -503,201 +394,6 @@ mod async_fd_shim {
                 Err(err) if err.kind() == io::ErrorKind::WouldBlock => Err(()),
                 Err(err) => Ok(Err(err)),
             }
-        }
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use core::{
-        convert::TryInto,
-        task::{Context, Poll},
-    };
-    use s2n_quic_core::{
-        endpoint::{self, CloseError},
-        event,
-        inet::SocketAddress,
-        io::{rx, tx},
-        path::Handle as _,
-        time::{Clock, Duration, Timestamp},
-    };
-    use std::collections::BTreeMap;
-
-    struct TestEndpoint {
-        addr: SocketAddress,
-        messages: BTreeMap<u32, Option<Timestamp>>,
-        now: Option<Timestamp>,
-        subscriber: NoopSubscriber,
-    }
-
-    impl TestEndpoint {
-        fn new(addr: SocketAddress) -> Self {
-            let messages = (0..1000).map(|id| (id, None)).collect();
-            Self {
-                addr,
-                messages,
-                now: None,
-                subscriber: Default::default(),
-            }
-        }
-    }
-
-    #[derive(Debug, Default)]
-    struct NoopSubscriber;
-
-    impl event::Subscriber for NoopSubscriber {
-        type ConnectionContext = ();
-
-        fn create_connection_context(
-            &mut self,
-            _meta: &event::api::ConnectionMeta,
-            _info: &event::api::ConnectionInfo,
-        ) -> Self::ConnectionContext {
-        }
-    }
-
-    impl Endpoint for TestEndpoint {
-        type PathHandle = PathHandle;
-        type Subscriber = NoopSubscriber;
-
-        const ENDPOINT_TYPE: endpoint::Type = endpoint::Type::Server;
-
-        fn transmit<Tx: tx::Queue<Handle = PathHandle>, C: Clock>(
-            &mut self,
-            queue: &mut Tx,
-            clock: &C,
-        ) {
-            let now = clock.get_time();
-            self.now = Some(now);
-
-            for (id, tx_time) in &mut self.messages {
-                match tx_time {
-                    Some(time)
-                        if now.saturating_duration_since(*time) < Duration::from_millis(50) =>
-                    {
-                        continue
-                    }
-                    _ => {
-                        let payload = id.to_be_bytes();
-                        let addr = PathHandle::from_remote_address(self.addr.into());
-                        let msg = (addr, payload);
-                        if queue.push(msg).is_ok() {
-                            *tx_time = Some(now);
-                        } else {
-                            // no more capacity
-                            return;
-                        }
-                    }
-                }
-            }
-        }
-
-        fn receive<Rx: rx::Queue<Handle = PathHandle>, C: Clock>(
-            &mut self,
-            queue: &mut Rx,
-            clock: &C,
-        ) {
-            let now = clock.get_time();
-            self.now = Some(now);
-
-            queue.for_each(|_header, payload| {
-                assert_eq!(payload.len(), 4, "invalid payload {:?}", payload);
-
-                let id = (&*payload).try_into().unwrap();
-                let id = u32::from_be_bytes(id);
-                self.messages.remove(&id);
-            });
-        }
-
-        fn poll_wakeups<C: Clock>(
-            &mut self,
-            _cx: &mut Context<'_>,
-            clock: &C,
-        ) -> Poll<Result<usize, CloseError>> {
-            let now = clock.get_time();
-            self.now = Some(now);
-
-            if self.messages.is_empty() {
-                return Err(CloseError).into();
-            }
-
-            Poll::Pending
-        }
-
-        fn timeout(&self) -> Option<Timestamp> {
-            self.now.map(|now| now + Duration::from_millis(50))
-        }
-
-        fn set_max_mtu(&mut self, _max_mtu: MaxMtu) {
-            // noop
-        }
-
-        fn subscriber(&mut self) -> &mut Self::Subscriber {
-            &mut self.subscriber
-        }
-    }
-
-    async fn test<A: std::net::ToSocketAddrs>(
-        receive_addr: A,
-        send_addr: Option<A>,
-    ) -> io::Result<()> {
-        let rx_socket = syscall::bind_udp(receive_addr, false)?;
-        let rx_socket: std::net::UdpSocket = rx_socket.into();
-        let addr = rx_socket.local_addr()?;
-
-        let mut io_builder = Io::builder().with_rx_socket(rx_socket)?;
-
-        if let Some(addr) = send_addr {
-            let tx_socket = syscall::bind_udp(addr, false)?;
-            let tx_socket: std::net::UdpSocket = tx_socket.into();
-            io_builder = io_builder.with_tx_socket(tx_socket)?
-        }
-
-        let io = io_builder.build()?;
-
-        let endpoint = TestEndpoint::new(addr.into());
-
-        let (task, local_addr) = io.start(endpoint)?;
-
-        let local_addr: std::net::SocketAddr = local_addr.into();
-
-        assert_eq!(local_addr, addr);
-
-        task.await?;
-
-        Ok(())
-    }
-
-    #[tokio::test]
-    async fn ipv4_test() -> io::Result<()> {
-        test("127.0.0.1:0", None).await
-    }
-
-    #[tokio::test]
-    async fn ipv4_two_socket_test() -> io::Result<()> {
-        test("127.0.0.1:0", Some("127.0.0.1:0")).await
-    }
-
-    #[tokio::test]
-    async fn ipv6_test() -> io::Result<()> {
-        match test(("::1", 0), None).await {
-            Err(err) if err.kind() == io::ErrorKind::AddrNotAvailable => {
-                eprintln!("The current environment does not support IPv6; skipping");
-                Ok(())
-            }
-            other => other,
-        }
-    }
-
-    #[tokio::test]
-    async fn ipv6_two_socket_test() -> io::Result<()> {
-        match test(("::1", 0), Some(("::1", 0))).await {
-            Err(err) if err.kind() == io::ErrorKind::AddrNotAvailable => {
-                eprintln!("The current environment does not support IPv6; skipping");
-                Ok(())
-            }
-            other => other,
         }
     }
 }
